Modified burst mode
Modified burst mode addresses the pump on and pump off scheme for the TA code where there are two pump on and 88 pump off pulses. This is used for samples that burn easily due to the repetition, also this mitigates the issue of slow buffer acquisition from slower repetition rates.